+++ netharn/metrics/drawing.py
@@ -195,7 +195,72 @@ def draw_perclass_prcurve(cx_to_peritem, classes=None, prefix='', fnum=1, **kw):
     return ax
 
 
-def draw_peritem_prcurve(peritem, prefix='', fnum=1, **kw):
+def draw_perclass_thresholds(cx_to_peritem, key='mcc', classes=None, prefix='', fnum=1, **kw):
+    """
+    Notes:
+        Each category is inspected independently of one another, there is no
+        notion of confusion.
+
+    Example:
+        >>> # xdoctest: +REQUIRES(module:ndsampler)
+        >>> # xdoctest: +REQUIRES(module:kwplot)
+        >>> from netharn.metrics.drawing import *  # NOQA
+        >>> from netharn.metrics import ConfusionVectors
+        >>> cfsn_vecs = ConfusionVectors.demo()
+        >>> classes = cfsn_vecs.classes
+        >>> ovr_cfsn = cfsn_vecs.binarize_ovr(keyby='name')
+        >>> cx_to_peritem = ovr_cfsn.threshold_curves()['perclass']
+        >>> import kwplot
+        >>> kwplot.autompl()
+        >>> key = 'mcc'
+        >>> draw_perclass_thresholds(cx_to_peritem, key, classes)
+        >>> # xdoctest: +REQUIRES(--show)
+        >>> kwplot.show_if_requested()
+    """
+    import kwplot
+    # Sort by descending "best value"
+    cxs = list(cx_to_peritem.keys())
+    priority = np.array([item['_max_' + key][0] for item in cx_to_peritem.values()])
+    priority[np.isnan(priority)] = -np.inf
+    cxs = list(ub.take(cxs, np.argsort(priority)))[::-1]
+    xydata = ub.odict()
+    for cx in cxs:
+        peritem = cx_to_peritem[cx]
+        if isinstance(cx, int):
+            catname = classes[cx]
+        else:
+            catname = cx
+
+        thresholds = peritem['thresholds']
+        measure = peritem[key]
+        best_label = peritem['max_{}'.format(key)]
+
+        nsupport = int(peritem['nsupport'])
+        if 'realpos_total' in peritem:
+            z = peritem['realpos_total']
+            if abs(z - int(z)) < 1e-8:
+                label = '{}: {} ({:d}/{:d})'.format(best_label, catname, int(peritem['realpos_total']), round(nsupport, 2))
+            else:
+                label = '{}: {} ({}/{:d})'.format(best_label, catname, round(peritem['realpos_total'], 2), round(nsupport, 2))
+        else:
+            label = '{}: {} ({:d})'.format(best_label, catname, round(nsupport, 2))
+        xydata[label] = (thresholds, measure)
+
+    with warnings.catch_warnings():
+        warnings.filterwarnings('ignore', 'Mean of empty slice', RuntimeWarning)
+
+    ax = kwplot.multi_plot(
+        xydata=xydata, fnum=fnum,
+        xlim=(0, 1), ylim=(0, 1), xpad=0.01, ypad=0.01,
+        xlabel='threshold', ylabel=key,
+        title=prefix + 'perclass {}'.format(key),
+        legend_loc='lower right',
+        color='distinct', linestyle='cycle', marker='cycle', **kw
+    )
+    return ax
+
+
+def draw_prcurve(peritem, prefix='', fnum=1, **kw):
     """
     TODO: rename to draw prcurve. Just draws a single pr curve.
 
@@ -211,7 +276,7 @@ def draw_peritem_prcurve(peritem, prefix='', fnum=1, **kw):
         >>> peritem = cfsn_vecs.binarize_peritem().precision_recall()
         >>> import kwplot
         >>> kwplot.autompl()
-        >>> draw_peritem_prcurve(peritem)
+        >>> draw_prcurve(peritem)
         >>> # xdoctest: +REQUIRES(--show)
         >>> kwplot.show_if_requested()
     """
@@ -253,4 +318,73 @@ def draw_peritem_prcurve(peritem, prefix='', fnum=1, **kw):
         legend_loc='lower right',
         color='distinct', linestyle='cycle', marker='cycle', **kw
     )
+    return ax
+
+
+def draw_threshold_curves(info, keys=None, prefix='', fnum=1, **kw):
+    """
+    Example:
+        >>> # xdoctest: +REQUIRES(module:ndsampler)
+        >>> # xdoctest: +REQUIRES(module:kwplot)
+        >>> import sys, ubelt
+        >>> sys.path.append(ubelt.expandpath('~/code/netharn'))
+        >>> from netharn.metrics.drawing import *  # NOQA
+        >>> from netharn.metrics import DetectionMetrics
+        >>> dmet = DetectionMetrics.demo(
+        >>>     nimgs=10, nboxes=(0, 10), n_fp=(0, 1), nclasses=3)
+        >>> cfsn_vecs = dmet.confusion_vectors()
+        >>> info = cfsn_vecs.binarize_peritem().threshold_curves()
+        >>> keys = None
+        >>> import kwplot
+        >>> kwplot.autompl()
+        >>> draw_threshold_curves(info, keys)
+        >>> # xdoctest: +REQUIRES(--show)
+        >>> kwplot.show_if_requested()
+    """
+    import kwplot
+    import kwimage
+    thresh = info['thresholds']
+
+    if keys is None:
+        keys = {'g1', 'f1', 'acc', 'mcc'}
+
+    idx_to_colors = kwimage.Color.distinct(len(keys), space='rgba')
+    idx_to_best_pt = {}
+
+    xydata = {}
+    colors = {}
+    for idx, key in enumerate(keys):
+        color = idx_to_colors[idx]
+        measure = info[key]
+        max_idx = measure.argmax()
+        best_thresh = thresh[max_idx]
+        best_measure = measure[max_idx]
+        best_label = '{}={:0.2f}@{:0.2f}'.format(key, best_measure, best_thresh)
+
+        nsupport = int(info['nsupport'])
+        if 'realpos_total' in info:
+            z = info['realpos_total']
+            if abs(z - int(z)) < 1e-8:
+                label = '{}: ({:d}/{:d})'.format(best_label, int(info['realpos_total']), round(nsupport, 2))
+            else:
+                label = '{}: ({}/{:d})'.format(best_label, round(info['realpos_total'], 2), round(nsupport, 2))
+        else:
+            label = '{}: ({:d})'.format(best_label, nsupport)
+        xydata[label] = (thresh, measure)
+        colors[label] = color
+        idx_to_best_pt[idx] = (best_thresh, best_measure)
+
+    ax = kwplot.multi_plot(
+        xydata=xydata, fnum=fnum,
+        xlim=(0, 1), ylim=(0, 1), xpad=0.01, ypad=0.01,
+        xlabel='threshold', ylabel=key,
+        title=prefix + 'info {}'.format(best_label),
+        legend_loc='lower right',
+        color=colors,
+        linestyle='cycle', marker='cycle', **kw
+    )
+    for idx, best_pt in idx_to_best_pt.items():
+        best_thresh, best_measure = best_pt
+        color = idx_to_colors[idx]
+        ax.plot(best_thresh, best_measure, '*', color=color)
     return ax